+++ src/app/services/sdk/presenters/index.tsx
@@ -10,7 +10,7 @@
 import {C_MSG} from '../const';
 import {AuthSentCode, AuthCheckedPhone, AuthAuthorization, AuthRecalled} from '../messages/chat.api.auth_pb';
 import {BlockedContactsMany, ContactsImported, ContactsMany} from '../messages/chat.api.contacts_pb';
-import {Bool, Error, Label, LabelsMany, MessageContainer} from '../messages/chat.core.types_pb';
+import {Bool, Error, Label, LabelsMany, MessageContainer, UserPhoto} from '../messages/chat.core.types_pb';
 import {MessagesDialogs, MessagesMany, MessagesSent} from '../messages/chat.api.messages_pb';
 import {Dialog, Group, GroupFull, PeerNotifySettings, User} from '../messages/chat.core.types_pb';
 import {
@@ -87,6 +87,8 @@ export default class Presenter {
                 return AccountPassword.deserializeBinary(data);
             case C_MSG.BlockedContactsMany:
                 return BlockedContactsMany.deserializeBinary(data);
+            case C_MSG.UserPhoto:
+                return UserPhoto.deserializeBinary(data);
             default:
                 return null;
         }